requirements.txt
openai
langchain
langchain-community
langchain-openai
langchainhub
langsmith
langchain-core
tabulate
azure-identity 
python-dotenv
azure-ai-documentintelligence
pypdf2
langchain
pypdf
python-docx
PyMuPDF
keras
numpy
pillow
tensorflow
streamlit
